+++ src/main/java/com/yuuko/core/commands/moderation/commands/BanCommand.java
@@ -34,13 +34,15 @@ public class BanCommand extends Command {
             return;
         }
 
+        if(!Sanitiser.canInteract(e, target, "ban", true)) {
+            return;
+        }
+
         if(commandParameters.length == 1) { // Case: Ban w/o reason and delDays.
             e.getGuild().getController().ban(target, 0).queue(s -> {
                 e.getMessage().addReaction("✅").queue();
                 ModerationLogSetting.execute(e, target.getUser(), 0, "None");
-            }, f -> {
-                e.getMessage().addReaction("❌").queue();
-            });
+            }, f -> e.getMessage().addReaction("❌").queue());
             return;
         }
 
@@ -50,16 +52,14 @@ public class BanCommand extends Command {
             delDays.set(Integer.parseInt(commandParameters[1]));
             if(delDays.get() > 7) {
                 delDays.set(7);
-                EmbedBuilder embed = new EmbedBuilder().setTitle("Invalid Input").setDescription("Deletion days must not be larger than 7. You input `**" + commandParameters[1] + "**`, so the actual value has been capped at 7.");
+                EmbedBuilder embed = new EmbedBuilder().setTitle("Invalid Input").setDescription("Deletion days must be no larger than 7. You input `**" + commandParameters[1] + "**`, so the actual value has been capped at 7.");
                 MessageHandler.sendMessage(e, embed.build());
             }
         } else {
             e.getGuild().getController().ban(target, 0, commandParameters[1]).queue(s -> { // Case: Ban w/reason, but no delDays.
                 e.getMessage().addReaction("✅").queue();
                 ModerationLogSetting.execute(e, target.getUser(), 0, commandParameters[1]);
-            }, f -> {
-                e.getMessage().addReaction("❌").queue();
-            });
+            }, f -> e.getMessage().addReaction("❌").queue());
             return;
         }
 
@@ -67,16 +67,12 @@ public class BanCommand extends Command {
             e.getGuild().getController().ban(target, delDays.get()).queue(s -> { // Case: Ban w/delDays, but no reason.
                 e.getMessage().addReaction("✅").queue();
                 ModerationLogSetting.execute(e, target.getUser(), delDays.get(), "None");
-            }, f -> {
-                e.getMessage().addReaction("❌").queue();
-            });
+            }, f -> e.getMessage().addReaction("❌").queue());
         } else {
             e.getGuild().getController().ban(target, delDays.get(), commandParameters[2]).queue(s -> { // Case: Ban w/reason and delDays.
                 e.getMessage().addReaction("✅").queue();
                 ModerationLogSetting.execute(e, target.getUser(), delDays.get(), commandParameters[2]);
-            }, f -> {
-                e.getMessage().addReaction("❌").queue();
-            });
+            }, f -> e.getMessage().addReaction("❌").queue());
         }
     }