+++ src/main/java/com/yuuko/core/commands/moderation/commands/MuteCommand.java
@@ -31,20 +31,20 @@ public class MuteCommand extends Command {
             return;
         }
 
+        if(!Sanitiser.canInteract(e, target, "mute", true)) {
+            return;
+        }
+
         if(target.getRoles().stream().noneMatch((role) -> role.getName().equalsIgnoreCase("Muted"))) {
             e.getGuild().getController().addSingleRoleToMember(target, Utils.setupMutedRole(e.getGuild())).queue(s -> {
                 e.getMessage().addReaction("✅").queue();
                 ModerationLogSetting.execute(e, "Mute", target.getUser(), (commandParameters.length < 2) ? "None" : commandParameters[1]);
-            }, f -> {
-                e.getMessage().addReaction("❌").queue();
-            });
+            }, f -> e.getMessage().addReaction("❌").queue());
         } else {
             e.getGuild().getController().removeSingleRoleFromMember(target, Utils.setupMutedRole(e.getGuild())).queue(s -> {
                 e.getMessage().addReaction("✅").queue();
                 ModerationLogSetting.execute(e, "Unmute", target.getUser(), (commandParameters.length < 2) ? "None" : commandParameters[1]);
-            }, f -> {
-                e.getMessage().addReaction("❌").queue();
-            });
+            }, f -> e.getMessage().addReaction("❌").queue());
         }
     }